+++ packages/ethereum-contracts/contracts/interfaces/superfluid/ISuperTokenFactory.sol
@@ -1,9 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >= 0.8.11;
 
+import { IERC20Metadata } from "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
 import { ISuperToken } from "./ISuperToken.sol";
-import { IERC20, ERC20WithTokenInfo } from "../tokens/ERC20WithTokenInfo.sol";
-
 /**
  * @title Super token factory interface
  * @author Superfluid
@@ -17,7 +16,7 @@ interface ISuperTokenFactory {
     error SUPER_TOKEN_FACTORY_DOES_NOT_EXIST();                 // 0x872cac48
     error SUPER_TOKEN_FACTORY_UNINITIALIZED();                  // 0x1b39b9b4
     error SUPER_TOKEN_FACTORY_ONLY_HOST();                      // 0x478b8e83
-    error SUPER_TOKEN_FACTORY_NON_UPGRADEABLE_IS_DEPRECATED();  // 0x478b8e83
+    error SUPER_TOKEN_FACTORY_NON_UPGRADEABLE_IS_DEPRECATED();  // 0xc4901a43
     error SUPER_TOKEN_FACTORY_ZERO_ADDRESS();                   // 0x305c9e82
 
     /**************************************************************************
@@ -59,7 +58,7 @@ interface ISuperTokenFactory {
      * @return superToken The deployed and initialized wrapper super token
      */
     function createERC20Wrapper(
-        IERC20 underlyingToken,
+        IERC20Metadata underlyingToken,
         uint8 underlyingDecimals,
         Upgradability upgradability,
         string calldata name,
@@ -79,7 +78,7 @@ interface ISuperTokenFactory {
      * - It assumes token provide the .decimals() function
      */
     function createERC20Wrapper(
-        ERC20WithTokenInfo underlyingToken,
+        IERC20Metadata underlyingToken,
         Upgradability upgradability,
         string calldata name,
         string calldata symbol
@@ -93,7 +92,7 @@ interface ISuperTokenFactory {
      * @param _underlyingToken Underlying ERC20 token
      * @return ISuperToken the created supertoken
      */
-    function createCanonicalERC20Wrapper(ERC20WithTokenInfo _underlyingToken)
+    function createCanonicalERC20Wrapper(IERC20Metadata _underlyingToken)
         external
         returns (ISuperToken);