+++ packages/ethereum-contracts/ops-scripts/deploy-framework.js
@@ -189,6 +189,7 @@ module.exports = eval(`(${S.toString()})({skipArgv: true})`)(async function (
         "Ownable",
         "CFAv1Forwarder",
         "IMultiSigWallet",
+        "ISafe",
         "SuperfluidGovernanceBase",
         "Resolver",
         "SuperfluidLoader",
@@ -204,6 +205,7 @@ module.exports = eval(`(${S.toString()})({skipArgv: true})`)(async function (
         "InstantDistributionAgreementV1",
         "ConstantOutflowNFT",
         "ConstantInflowNFT",
+        "IAccessControlEnumerable",
     ];
     const mockContracts = [
         "SuperfluidMock",
@@ -213,6 +215,7 @@ module.exports = eval(`(${S.toString()})({skipArgv: true})`)(async function (
     const {
         Ownable,
         IMultiSigWallet,
+        ISafe,
         CFAv1Forwarder,
         SuperfluidGovernanceBase,
         Resolver,
@@ -359,8 +362,8 @@ module.exports = eval(`(${S.toString()})({skipArgv: true})`)(async function (
     if (governanceInitializationRequired) {
         const accounts = await web3.eth.getAccounts();
         const trustedForwarders = [];
-        if (config.biconomyForwarder) {
-            trustedForwarders.push(config.biconomyForwarder);
+        if (config.trustedForwarders) {
+            trustedForwarders.push(...config.trustedForwarders);
         }
         if (config.cfaFwd) {
             trustedForwarders.push(config.cfaFwd);
@@ -866,6 +869,7 @@ module.exports = eval(`(${S.toString()})({skipArgv: true})`)(async function (
                 contracts: {
                     Ownable,
                     IMultiSigWallet,
+                    ISafe,
                     SuperfluidGovernanceBase,
                 },
             },
@@ -885,6 +889,8 @@ module.exports = eval(`(${S.toString()})({skipArgv: true})`)(async function (
         const sfObjForResolver = {
             contracts: {
                 Resolver,
+                IMultiSigWallet,
+                ISafe,
                 IAccessControlEnumerable,
             },
             resolver: {