+++ packages/ethereum-contracts/ops-scripts/gov-create-new-app-registration-key.js
@@ -56,6 +56,7 @@ module.exports = eval(`(${S.toString()})({
         additionalContracts: [
             "Ownable",
             "IMultiSigWallet",
+            "ISafe",
             "SuperfluidGovernanceBase",
         ],
         contractLoader: builtTruffleContractLoader,