+++ packages/ethereum-contracts/ops-scripts/gov-set-token-min-deposit.js
@@ -37,6 +37,7 @@ module.exports = eval(`(${S.toString()})()`)(async function (
         additionalContracts: [
             "Ownable",
             "IMultiSigWallet",
+            "ISafe",
             "SuperfluidGovernanceBase",
         ],
         contractLoader: builtTruffleContractLoader,