+++ packages/ethereum-contracts/ops-scripts/gov-upgrade-governance.js
@@ -34,6 +34,7 @@ module.exports = eval(`(${S.toString()})()`)(async function (
         additionalContracts: [
             "Ownable",
             "IMultiSigWallet",
+            "ISafe",
             "SuperfluidGovernanceBase",
             "SuperfluidGovernanceII",
         ],