+++ packages/ethereum-contracts/ops-scripts/resolver-list-super-token.js
@@ -39,6 +39,7 @@ module.exports = eval(`(${S.toString()})()`)(async function (
         additionalContracts: [
             "Ownable",
             "IMultiSigWallet",
+            "ISafe",
             "SuperfluidGovernanceBase",
             "SuperToken",
             "Resolver",