+++ packages/sdk-core/src/SuperToken.ts
@@ -1,9 +1,3 @@
-import {
-    ISETH,
-    ISETH__factory,
-    ISuperToken,
-    ISuperToken__factory,
-} from "@superfluid-finance/ethereum-contracts/build/typechain-ethers-v5";
 import { BytesLike, ethers, Overrides } from "ethers";
 
 import ConstantFlowAgreementV1 from "./ConstantFlowAgreementV1";
@@ -50,6 +44,12 @@ import {
     SuperTokenFlowRateAllowanceParams,
     SuperTokenFlowRateAllowanceWithPermissionsParams,
 } from "./interfaces";
+import {
+    ISETH,
+    ISETH__factory,
+    ISuperToken,
+    ISuperToken__factory,
+} from "./typechain-types";
 import {
     getSanitizedTimestamp,
     getStringCurrentTimeInSeconds,