+++ packages/ethereum-contracts/CHANGELOG.md
@@ -5,6 +5,29 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Unreleased
 
+### Breaking
+- `TokenInfo` and `ERC20WithTokenInfo` interface/abstract contract are removed from the codebase, including the bundled ABI contracts
+  - Migration: Use `IERC20Metadata` instead, as this replaces the previous contracts
+- `build/typechain-ethers-v5` is removed from the npm package
+  - Migration: Consume the ABIs and utilize whatever tool you'd like to generate types from them
+
+### Added
+
+- Added 'test-slither' yarn sub-task.
+- Expose `SuperToken._underlyingDecimals` with `SuperToken.getUnderlyingDecimals()`
+- Expose `_toUnderlyingAmount(uint256 amount)` with `toUnderlyingAmount(uint256 amount)`
+- `batchCall` supports payable `OPERATION_TYPE_SUPERFLUID_CALL_APP_ACTION`: only the first `OPERATION_TYPE_SUPERFLUID_CALL_APP_ACTION` will be payable
+
+### Changed
+
+## [v1.8.1] - 2023-08-28
+
+### Fixed
+
+- Publish `build/hardhat/*/*` again.
+
+## [v1.8.0] - 2023-08-23
+
 ### Added
 
 - Publish `build/bundled-abi.{js,json}` in the npm package.