+++ packages/metadata/CHANGELOG.md
@@ -3,6 +3,23 @@ All notable changes to the metadata will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.1.16]
+### Fixed
+- Fixed `gdaV1` address for `avalanche-fuji`
+
+## [v1.1.15]
+### Fixed
+- Fixed `existentialNFTCloneFactory` address for `celo-mainnet`
+
+## [v1.1.14]
+
+### Added
+- Added `constantOutflowNFT` and `constantInflowNFT`
+
+## [v1.1.13]
+### Added
+- Added field `existentialNFTCloneFactory` to contract addresses
+
 ## [v1.1.12]
 ### Added
 - Added subgraph endpoints for: Autowrap, FlowScheduler and Vesting contracts